Refactor: RectangleViewFactory 생성하여 ViewController에서 RectangleView생성하는 부분 삭제
